+++ src/dev.rs
@@ -26,7 +26,7 @@ pub fn discover() -> Result<RawIter> {
 	let mut len = 0;
 
 	let n = unsafe { ffi::LIBMTP_Detect_Raw_Devices(&mut ptr, &mut len) };
-	match err::Kind::from_number(n) {
+	match err::Kind::from_ffi(n) {
 		Some(kind) if kind != err::Kind::NoDeviceAttached => {
 			Err(Error::new(kind, "Failed to discover raw devices"))
 		}
@@ -38,31 +38,28 @@ pub fn discover() -> Result<RawIter> {
 #[derive(Clone, Hash)]
 pub struct Device {
 	/// The underlying structure of the device.
-	inner: *mut ffi::LIBMTP_mtpdevice_t,
+	inner: ffi::LIBMTP_mtpdevice_t,
+	/// The pointer to the underlying structure of the device.
+	inner_ptr: *mut ffi::LIBMTP_mtpdevice_t,
 }
 
 impl Device {
-	/// Constructs a device from the underlying structure.
-	fn new(inner: *mut ffi::LIBMTP_mtpdevice_t) -> Self {
-		Self { inner }
+	/// Constructs a new device.
+	pub(crate) fn from_ffi(ptr: *mut ffi::LIBMTP_mtpdevice_t) -> Self {
+		Self { inner: unsafe { *ptr }, inner_ptr: ptr }
 	}
 
-	/// Retieves the friendly name of the device.
+	/// Refreshes storages information for the device.
 	///
-	/// # Errors
-	///
-	/// Returns an error if the device doesn't have a support for friendly names or if the friendly name was not found.
-	pub fn name(&self) -> Result<String> {
-		let ptr = unsafe { ffi::LIBMTP_Get_Friendlyname(self.inner) };
-		if ptr.is_null() {
+	/// Call this function before displaying storage information.
+	pub fn refresh(&mut self) -> Result<()> {
+		let res = unsafe {
+			ffi::LIBMTP_Get_Storage(self.inner_ptr, ffi::LIBMTP_STORAGE_SORTBY_NOTSORTED)
+		};
+		if res != 0 {
 			return Err(self.pop_err().unwrap_or_default());
 		}
-
-		let name = unsafe { ptr_to_string(ptr) };
-		unsafe {
-			libc::free(ptr as *mut c_void);
-		}
-		Ok(name)
+		Ok(())
 	}
 
 	/// Changes the friendly name of the device.
@@ -74,89 +71,117 @@ impl Device {
 	/// # Panics
 	///
 	/// Panics if `name` contains a nul byte.
-	pub fn rename(&mut self, name: &str) -> Result<()> {
+	pub fn rename(&self, name: &str) -> Result<()> {
 		let name = CString::new(name).expect("Name should not contain a nul byte");
-		let n = unsafe { ffi::LIBMTP_Set_Friendlyname(self.inner, name.as_ptr()) };
+		let n = unsafe { ffi::LIBMTP_Set_Friendlyname(self.inner_ptr, name.as_ptr()) };
 		if n != 0 {
 			return Err(self.pop_err().unwrap_or_default());
 		}
 		Ok(())
 	}
 
+	/// Retieves the friendly name of the device.
+	///
+	/// # Errors
+	///
+	/// Returns an error if the device doesn't have a support for friendly names or if the friendly name was not found.
+	pub fn name(&self) -> Result<String> {
+		let ptr = unsafe { ffi::LIBMTP_Get_Friendlyname(self.inner_ptr) };
+		if ptr.is_null() {
+			return Err(self.pop_err().unwrap_or_default());
+		}
+
+		let name = unsafe { ptr_to_string(ptr) };
+		unsafe {
+			libc::free(ptr as *mut c_void);
+		}
+		Ok(name)
+	}
+
 	/// Retrieves the maximum battery percentage of the device.
 	pub fn max_battery_percent(&self) -> u8 {
-		(unsafe { *self.inner }).maximum_battery_level
+		self.inner.maximum_battery_level
 	}
 
 	/// Retrieves the ID of the default music folder of the device.
 	///
 	/// If the default music folder was not found, the ID of the root folder will be returned.
 	pub fn music_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_music_folder
+		self.inner.default_music_folder
 	}
 
 	/// Retrieves the ID of the default playlists folder of the device.
 	///
 	/// If the default playlists folder was not found, the ID of the root folder will be returned.
 	pub fn playlist_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_playlist_folder
+		self.inner.default_playlist_folder
 	}
 
 	/// Retrieves the ID of the default pictures folder of the device.
 	///
 	/// If the default pictures folder was not found, the ID of the root folder will be returned.
 	pub fn picture_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_picture_folder
+		self.inner.default_picture_folder
 	}
 
 	/// Retrieves the ID of the default videos folder of the device.
 	///
 	/// If the default videos folder was not found, the ID of the root folder will be returned.
 	pub fn video_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_video_folder
+		self.inner.default_video_folder
 	}
 
 	/// Retrieves the ID of the default organizers folder of the device.
 	///
 	/// If the default organizers folder was not found, the ID of the root folder will be returned.
 	pub fn organizer_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_organizer_folder
+		self.inner.default_organizer_folder
 	}
 
 	/// Retrieves the ID of the default ZENcast folder of the device.
 	///
 	/// If the default ZENcast folder was not found, the ID of the root folder will be returned.
 	pub fn zencast_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_zencast_folder
+		self.inner.default_zencast_folder
 	}
 
 	/// Retrieves the ID of the default albums folder of the device.
 	///
 	/// If the default albums folder was not found, the ID of the root folder will be returned.
 	pub fn album_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_album_folder
+		self.inner.default_album_folder
 	}
 
 	/// Retrieves the ID of the default texts folder of the device.
 	///
 	/// If the default texts folder was not found, the ID of the root folder will be returned.
 	pub fn text_folder_id(&self) -> u32 {
-		(unsafe { *self.inner }).default_text_folder
+		self.inner.default_text_folder
 	}
 
 	/// Retrieves an iterator over the storages of the device.
-	pub fn storages(&self) -> storage::Iter {
-		storage::Iter::new(self, (unsafe { *self.inner }).storage)
+	pub fn iter(&self) -> storage::Iter {
+		storage::Iter::new(self)
+	}
+
+	/// Retrieves the underlying structure of the device.
+	pub(crate) fn inner(&self) -> ffi::LIBMTP_mtpdevice_t {
+		self.inner
+	}
+
+	/// Retrieves the pointer to the underlying structure of the device.
+	pub(crate) fn inner_ptr(&self) -> *mut ffi::LIBMTP_mtpdevice_t {
+		self.inner_ptr
 	}
 
 	/// Pops the last error from the error stack.
 	///
 	/// After the execution the error stack will be cleared.
 	fn pop_err(&self) -> Option<Error> {
-		let stack = unsafe { ffi::LIBMTP_Get_Errorstack(self.inner) };
-		let err = Error::from_stack(stack);
+		let stack = unsafe { ffi::LIBMTP_Get_Errorstack(self.inner_ptr) };
+		let err = Error::from_ffi(stack);
 		unsafe {
-			ffi::LIBMTP_Clear_Errorstack(self.inner);
+			ffi::LIBMTP_Clear_Errorstack(self.inner_ptr);
 		}
 		err
 	}
@@ -166,7 +191,7 @@ impl Device {
 impl Drop for Device {
 	fn drop(&mut self) {
 		unsafe {
-			ffi::LIBMTP_Release_Device(self.inner);
+			ffi::LIBMTP_Release_Device(self.inner_ptr);
 		}
 	}
 }
@@ -192,20 +217,35 @@ impl<'a> IntoIterator for &'a Device {
 	type IntoIter = storage::Iter<'a>;
 
 	fn into_iter(self) -> Self::IntoIter {
-		self.storages()
+		self.iter()
 	}
 }
 
 /// A device connected via USB, but not yet opened.
+#[derive(Clone, Hash)]
 pub struct RawDevice {
 	/// The underlying structure of the device.
 	inner: ffi::LIBMTP_raw_device_t,
+	/// The pointer to the underlying structure of the device.
+	inner_ptr: *mut ffi::LIBMTP_raw_device_t,
 }
 
 impl RawDevice {
-	/// Constructs a raw device from the underlying structure.
-	fn new(inner: ffi::LIBMTP_raw_device_t) -> Self {
-		Self { inner }
+	/// Constructs a new raw device.
+	pub(crate) fn from_ffi(ptr: *mut ffi::LIBMTP_raw_device_t) -> Self {
+		Self { inner: unsafe { *ptr }, inner_ptr: ptr }
+	}
+
+	/// Attempts to open the device, with caching.
+	pub fn open(&self) -> Option<Device> {
+		let ptr = unsafe { ffi::LIBMTP_Open_Raw_Device(self.inner_ptr) };
+		if ptr.is_null() { None } else { Some(Device::from_ffi(ptr)) }
+	}
+
+	/// Attempts to open the device, without caching.
+	pub fn open_uncached(&self) -> Option<Device> {
+		let ptr = unsafe { ffi::LIBMTP_Open_Raw_Device_Uncached(self.inner_ptr) };
+		if ptr.is_null() { None } else { Some(Device::from_ffi(ptr)) }
 	}
 
 	/// Retrieves the vendor of the device.
@@ -214,13 +254,7 @@ impl RawDevice {
 	///
 	/// Panics if the vendor name of the device is not a valid UTF-8.
 	pub fn vendor(&self) -> Vendor {
-		let id = self.inner.device_entry.vendor_id;
-		let name = unsafe {
-			CStr::from_ptr(self.inner.device_entry.vendor)
-				.to_str()
-				.expect("Vendor name should be a valid UTF-8")
-		};
-		Vendor::new(id, name)
+		Vendor::from_ffi(self.inner.device_entry)
 	}
 
 	/// Retrieves the product of the device.
@@ -229,27 +263,16 @@ impl RawDevice {
 	///
 	/// Panics if the product name of the device is not a valid UTF-8.
 	pub fn product(&self) -> Product {
-		let id = self.inner.device_entry.product_id;
-		let name = unsafe {
-			CStr::from_ptr(self.inner.device_entry.product)
-				.to_str()
-				.expect("Product name should be a valid UTF-8")
-		};
-		Product::new(id, name)
+		Product::from_ffi(self.inner.device_entry)
 	}
+}
 
-	/// Attempts to open the device, with caching.
-	pub fn open(&self) -> Option<Device> {
-		let ptr = &self.inner as *const _;
-		let dev_ptr = unsafe { ffi::LIBMTP_Open_Raw_Device(ptr as *mut _) };
-		if dev_ptr.is_null() { None } else { Some(Device::new(dev_ptr)) }
-	}
-
-	/// Attempts to open the device, without caching.
-	pub fn open_uncached(&self) -> Option<Device> {
-		let ptr = &self.inner as *const _;
-		let dev_ptr = unsafe { ffi::LIBMTP_Open_Raw_Device_Uncached(ptr as *mut _) };
-		if dev_ptr.is_null() { None } else { Some(Device::new(dev_ptr)) }
+#[doc(hidden)]
+impl Drop for RawDevice {
+	fn drop(&mut self) {
+		unsafe {
+			libc::free(self.inner_ptr as *mut c_void);
+		}
 	}
 }
 
@@ -263,18 +286,36 @@ impl Debug for RawDevice {
 }
 
 /// A vendor of the device.
-#[derive(Copy, Clone, Hash, Debug)]
+#[derive(Clone, Copy, Hash, Debug)]
 pub struct Vendor<'a> {
 	/// The ID of the vendor.
-	pub id: u16,
+	id: u16,
 	/// The name of the vendor.
-	pub name: &'a str,
+	name: &'a str,
 }
 
 impl<'a> Vendor<'a> {
-	/// Constructs a new vendor from the id and the name.
-	fn new(id: u16, name: &'a str) -> Self {
+	/// Constructs a new vendor.
+	///
+	/// # Panics
+	///
+	/// Panics if the name of the vendor is not a valid UTF-8.
+	pub(crate) fn from_ffi(inner: ffi::LIBMTP_device_entry_t) -> Self {
+		let id = inner.vendor_id;
+		let name = unsafe {
+			CStr::from_ptr(inner.vendor).to_str().expect("Vendor name should be a valid UTF-8")
+		};
 		Self { id, name }
+	}
+
+	/// Retrieves the ID of the vendor.
+	pub fn id(&self) -> u16 {
+		self.id
+	}
+
+	/// Retrieves the name of the vendor.
+	pub fn name(&self) -> &str {
+		self.name
 	}
 }
 
@@ -285,18 +326,36 @@ impl<'a> Display for Vendor<'a> {
 }
 
 /// A product of the device.
-#[derive(Copy, Clone, Hash, Debug)]
+#[derive(Clone, Copy, Hash, Debug)]
 pub struct Product<'a> {
 	/// The ID of the product.
-	pub id: u16,
+	id: u16,
 	/// The name of the product.
-	pub name: &'a str,
+	name: &'a str,
 }
 
 impl<'a> Product<'a> {
-	/// Constructs a new product from the id and the name.
-	fn new(id: u16, name: &'a str) -> Self {
+	/// Constructs a new product.
+	///
+	/// # Panics
+	///
+	/// Panics if the name of the product is not a valid UTF-8.
+	pub(crate) fn from_ffi(inner: ffi::LIBMTP_device_entry_t) -> Self {
+		let id = inner.product_id;
+		let name = unsafe {
+			CStr::from_ptr(inner.product).to_str().expect("Product name should be a valid UTF-8")
+		};
 		Self { id, name }
+	}
+
+	/// Retrieves the ID of the product.
+	pub fn id(&self) -> u16 {
+		self.id
+	}
+
+	/// Retrieves the name of the product.
+	pub fn name(&self) -> &str {
+		self.name
 	}
 }
 
@@ -307,6 +366,7 @@ impl<'a> Display for Product<'a> {
 }
 
 /// An iterator over the raw devices.
+#[derive(Clone, Copy)]
 pub struct RawIter {
 	ptr: *mut ffi::LIBMTP_raw_device_t,
 	len: isize,
@@ -314,8 +374,8 @@ pub struct RawIter {
 }
 
 impl RawIter {
-	/// Constructs a new raw device iterator from the underlying structure.
-	fn new(ptr: *mut ffi::LIBMTP_raw_device_t, len: isize) -> Self {
+	/// Constructs a new raw device iterator.
+	pub(crate) fn new(ptr: *mut ffi::LIBMTP_raw_device_t, len: isize) -> Self {
 		Self { ptr, len, off: 0 }
 	}
 }
@@ -328,18 +388,9 @@ impl Iterator for RawIter {
 			return None;
 		}
 
-		let ptr = unsafe { *self.ptr.offset(self.off) };
-		let dev = RawDevice::new(ptr);
+		let ptr = unsafe { self.ptr.offset(self.off) };
+		let dev = RawDevice::from_ffi(ptr);
 		self.off += 1;
 		Some(dev)
-	}
-}
-
-#[doc(hidden)]
-impl Drop for RawIter {
-	fn drop(&mut self) {
-		unsafe {
-			libc::free(self.ptr as *mut c_void);
-		}
 	}
 }